Styling changes and refactoring
