+++ src/components/chat/components/ChatMessage.js
@@ -1,86 +1,26 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { Button, Dropdown, DropdownMenu } from '@folio/stripes/components';
-import moment from 'moment';
+import { Button, Dropdown, DropdownMenu, IconButton } from '@folio/stripes/components';
 
 import classNames from 'classnames';
+
+import ChatMessageHeader from './ChatMessageHeader';
 import css from './ChatMessage.css';
+import MessageDropdown from './MessageDropdown';
+
+const systemMessageKeys = [
+  '#ReShareLoanConditionAgreeResponse#',
+  '#ReShareSupplierConditionsAssumedAgreed#',
+  '#ReShareSupplierAwaitingConditionConfirmation#',
+  '#ReShareAddLoanCondition#'
+];
 
 const ChatMessage = React.forwardRef((props, ref) => {
   const { notification } = props;
 
   const lowerCaseFirstLetter = (string) => {
     return string.charAt(0).toLowerCase() + string.slice(1);
-  };
-
-  const systemMessageKeys = [
-    '#ReShareLoanConditionAgreeResponse#',
-    '#ReShareSupplierConditionsAssumedAgreed#',
-    '#ReShareSupplierAwaitingConditionConfirmation#',
-    '#ReShareAddLoanCondition#'
-  ];
-
-  const longDateFormatter = (timestamp) => {
-    // This takes a moment timestamp
-    return `${timestamp.format('MMM. D, YYYY ')} at ${timestamp.format(' h:mm a')}`;
-  };
-
-  const renderDateTime = () => {
-    const duration = moment.duration(moment().diff(moment(notification?.timestamp)));
-
-    const { days, hours, minutes, seconds } = duration?._data || {};
-    const JUST_NOW_THRESHOLD = 30;
-    // If more than a day, return long date formatted
-    if (days !== 0) {
-      return (longDateFormatter(moment(notification?.timestamp)));
-    }
-
-    // If time in hours
-    if (hours > 1) {
-      return <FormattedMessage id="ui-rs.view.chatMessage.hours" values={{ hours }} />;
-    }
-    if (hours === 1) {
-      return <FormattedMessage id="ui-rs.view.chatMessage.hour" />;
-    }
-
-    if (minutes > 1) {
-      return <FormattedMessage id="ui-rs.view.chatMessage.minutes" values={{ minutes }} />;
-    }
-
-    if (minutes === 1) {
-      return <FormattedMessage id="ui-rs.view.chatMessage.minute" />;
-    }
-
-    if (seconds > JUST_NOW_THRESHOLD) {
-      return <FormattedMessage id="ui-rs.view.chatMessage.seconds" values={{ seconds }} />;
-    }
-
-    return <FormattedMessage id="ui-rs.view.chatMessage.justNow" />; 
-  };
-
-  const renderHeader = () => {
-    return (
-      <div
-        className={css.header}
-      >
-        <b>
-          {notification?.messageSender?.owner?.name}
-        </b>
-        <span className={css.headerTime}>&nbsp;</span>
-        <span className={css.headerTime}>
-          {renderDateTime()}
-        </span>
-        {!notification?.isSender &&
-          <>
-            <span className={css.headerTime}>&nbsp; &#183; &nbsp;</span>
-            <span className={css.headerTime}>
-              {notification?.seen ? <FormattedMessage id="ui-rs.view.chatMessage.actions.Read" /> : <b><FormattedMessage id="ui-rs.view.chatMessage.actions.Unread" /></b>}
-            </span>
-          </>
-        }
-      </div>
-    );
   };
 
   const renderActionContents = () => {
@@ -113,47 +53,6 @@ const ChatMessage = React.forwardRef((props, ref) => {
         </span>
         <span>&nbsp;</span>
       </>
-    );
-  };
-
-  const renderDropdownButtonContents = () => {
-    const { onToggle } = props;
-    const markAsReadText = notification?.seen ?
-      <FormattedMessage id="ui-rs.view.chatMessage.actions.markAsUnread" /> :
-      <FormattedMessage id="ui-rs.view.chatMessage.actions.markAsRead" />;
-
-    return (
-      <DropdownMenu
-        data-role="menu"
-        aria-label="actions-for-message"
-        onToggle={onToggle}
-      >
-        <FormattedMessage id="ui-rs.view.chatMessage.actions">
-          {ariaLabel => (
-            !notification?.isSender ?
-              <Button
-                aria-label={ariaLabel}
-                buttonStyle="dropdownItem"
-                id="clickable-mark-message-read"
-                marginBottom0
-                onClick={() => props.handleMessageRead(notification, notification.seen)}
-              >
-                {markAsReadText}
-              </Button> : <FormattedMessage id="ui-rs.view.chatMessage.actions.noAvailableActions" />
-          )}
-        </FormattedMessage>
-      </DropdownMenu>
-    );
-  };
-
-  const renderDropdownButton = () => {
-    return (
-      <Dropdown
-        buttonProps={{ marginBottom0: true }}
-        className={css.actionButton}
-        label={<FormattedMessage id="ui-rs.view.chatMessage.actions" />}
-        renderMenu={renderDropdownButtonContents}
-      />
     );
   };
 
@@ -197,6 +96,9 @@ const ChatMessage = React.forwardRef((props, ref) => {
   };
 
   const className = classOfMessageCard();
+  const markAsReadText = notification?.seen ?
+    <FormattedMessage id="ui-rs.view.chatMessage.actions.markAsUnread" /> :
+    <FormattedMessage id="ui-rs.view.chatMessage.actions.markAsRead" />;
   return (
     <>
       <div
@@ -210,11 +112,20 @@ const ChatMessage = React.forwardRef((props, ref) => {
           className={classNames(className, css.displayFlex)}
         >
           <div className={css.leftHandSide}>
-            {renderHeader()}
+            <ChatMessageHeader notification={notification} />
             {renderContents()}
           </div>
           <div className={css.rightHandSide}>
-            {renderDropdownButton()}
+            {!notification?.isSender &&
+              <MessageDropdown
+                actionItems={[
+                  {
+                    label: markAsReadText,
+                    onClick: () => props.handleMessageRead(notification, notification.seen)
+                  }
+                ]}
+              />
+            }
           </div>
         </div>
       </div>
@@ -222,11 +133,7 @@ const ChatMessage = React.forwardRef((props, ref) => {
   );
 });
 
-/* eslint-disable react/no-unused-prop-types */
-// For some reason eslint complains when the onToggle prop is here AND when it isn't, so I'm putting it here to be safe, and shushing lint.
-
 ChatMessage.propTypes = {
-  handleMessageRead: PropTypes.func.isRequired,
   notification: PropTypes.shape({
     id: PropTypes.string,
     messageContent: PropTypes.string,
@@ -248,8 +155,7 @@ ChatMessage.propTypes = {
         name: PropTypes.string,
       })
     }),
-  }),
-  onToggle: PropTypes.func.isRequired,
+  })
 };
 
 export default ChatMessage;
